lots of changes to implement e-mail verification
tests need to be touched up and CI might break. tests were also altered so that CI won't break quite so bad if it does

added dotenv to the project in order to safely use the SMTP server. changed docker-compose to accommodate the new .env file

added a new `verifyUser` function to create the `api/auth/verify/:token` route

the frontend is likely broken now that signup is a two step process. will need to change some logic to check if the user is verified before showing the homepage